+++ tests/fixtures/get-sample-project.ts
@@ -1,0 +1,34 @@
+import execa from "execa"
+import { build } from "lib/build"
+import path from "path"
+
+export const getSampleProject = async () => {
+  const projectPath = path.resolve(__dirname, "assets", "nextjs-sample-project")
+
+  await execa("yarn", [], { cwd: projectPath })
+
+  await build({
+    dir: projectPath
+  })
+
+  const nsmIndex = require(path.resolve(
+    __dirname,
+    "assets",
+    "nextjs-sample-project",
+    ".nsm",
+    "index.ts"
+  ))
+
+  const getServerFixture = require(path.resolve(
+    __dirname,
+    "assets",
+    "nextjs-sample-project",
+    ".nsm",
+    "get-server-fixture.ts"
+  ))
+
+  return {
+    nsmIndex: nsmIndex.default,
+    getServerFixture: getServerFixture.default
+  }
+}
